requirements.txt
absl-py==1.4.0
astunparse==1.6.3
atom3d==0.2.6
biopython==1.80
blosc2==2.0.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Cython==0.29.33
dill==0.3.6
easy-parallel==0.1.6
flatbuffers==23.1.4
freesasa==2.1.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
llvmlite==0.39.1
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
multipledispatch==0.6.0
multiprocess==0.70.14
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.2
pathos==0.3.0
pox==0.3.2
ppft==1.7.6.6
protobuf==3.19.6
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrr==0.10.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
SQLAlchemy==1.4.46
tables==3.8.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
torch==1.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
